pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentf"
version = "0.2.1"
description = "Self-hosted AI knowledge platform with 27 data source connectors, hybrid search, and MCP server"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "canmeng", email = "gcanmeng@gmail.com" }
]
keywords = [
    "rag", "knowledge-base", "mcp", "claude", "vector-search",
    "llm", "ai", "fastapi", "chromadb"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    # Vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    # Document parsing
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    # Auth & config
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "sqlalchemy>=2.0.0",
    # MCP & HTTP
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    # Scheduler
    "APScheduler>=3.10.0",
    # Object storage connectors
    "oss2>=2.18.0",
    "boto3>=1.34.0",
    "cos-python-sdk-v5>=1.9.0",
    "esdk-obs-python>=3.24.0",
    # Relational / OLAP connectors
    "PyMySQL>=1.1.0",
    "oracledb>=2.0.0",
    "pymssql>=2.2.0",
    "snowflake-sqlalchemy>=1.5.0",
    "clickhouse-connect>=0.7.0",
    "pyhive[hive]>=0.7.0",
    "thrift>=0.16.0",
    # Search / NoSQL connectors
    "elasticsearch>=8.0.0",
    "pymongo>=4.6.0",
    # Enterprise cloud connectors
    "msal>=1.24.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27.0",
    "build>=1.0",
    "twine>=5.0",
]

[project.urls]
Homepage = "https://github.com/canmengfly/AgentForge"
Documentation = "https://github.com/canmengfly/AgentForge/tree/main/docs"
Repository = "https://github.com/canmengfly/AgentForge"
"Bug Tracker" = "https://github.com/canmengfly/AgentForge/issues"

[project.scripts]
agentf-api = "src.api.main:run"
agentf-mcp = "src.mcp.server:run"

# ── Package discovery ─────────────────────────────────────────────────────────

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "*.egg-info*"]

[tool.setuptools.package-data]
"src" = [
    "templates/**/*",
    "static/**/*",
    "static/css/*",
    "static/js/*",
]
